Move readFile() into spec folder
