+++ lib/sftp.js
@@ -26,22 +26,6 @@ exports.close = function close(sftp, handle) {
   });
 };
 
-exports.readFile = function readFile(client, path, cb) {
-  const readStream = client.createReadStream(path);
-
-  const chunks = [];
-
-  readStream.on('data', (chunk) => {
-    chunks.push(chunk);
-  });
-
-  readStream.on('end', () => {
-    const buffer = Buffer.concat(chunks);
-
-    cb(null, buffer);
-  });
-};
-
 exports.createPath = function createPath(dir, file) {
   let slash = '';
